pyproject.toml
[project]
name = "exporter"
version = "0.1.0"
description = "Prometheus exporter for Supra blockchain RPC and Validator metrics"
authors = [
  { name = "Abhinav Taneja", email = "at@blocksize-capital.com" },
  { name = "Axel Lode", email = "al@blocksize-capital.com" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
  "prometheus-client>=0.21.0,<0.22.0",
  "python-dotenv>=1.0.0,<2.0.0",
  "requests>=2.31.0,<3.0.0",
]

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
